+++ src/main/java/chess/controller/ChessController.java
@@ -2,7 +2,7 @@ package chess.controller;
 
 import java.util.NoSuchElementException;
 
-import chess.converter.console.StringToCommandConverter;
+import chess.controller.converter.StringToCommandConverter;
 import chess.controller.dto.BoardDto;
 import chess.domain.board.ChessScore;
 import chess.domain.board.BoardInitializer;
